+++ airfs/storage/github/_model_archive.py
@@ -22,7 +22,7 @@ class Archive(GithubObject):
             spec (dict): Item spec.
 
         Returns:
-            generator of tuple: object name str, object header dict, directory bool
+            generator of tuple: object name str, object header dict, has content bool
         """
         cls._raise_if_not_dir(spec.get("archive"), spec)